pyproject.toml
[project]
name = "nafdac"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib==2.214.0",
    "constructs>=10.0.0,<11.0.0",
    "scrapy>=2.14.0",
    "selenium>=4.39.0",
    "aws-lambda-powertools>=3.3.0",
]

[dependency-groups]
dev = [
    "pytest==6.2.5",
    "boto3>=1.26.0",
    "moto>=4.0.0",
]
